Changed indexes to uint16
There is only ~130000 cities to choose from which could be mapped with only 9 bits so 16 bits per grid direction should be enough for realistic options for the simulate.

This results in a massive reduction in the memory required to store the maps, and Go transparently permits any integer type to be used as array indices.